requirements.txt
openai>=1.69.0
python-dotenv>=1.1.0
pydantic>=2.10.6
fastapi>=1.112.0
uvicorn>=0.34.0
python-multipart>=0.0.20
langchain>=0.3.21
langchain-openai>=0.3.11
supabase==2.15.0
pytest>=8.3.5